SISRP-45434 - Post-3/3 React Cleanup - move CampusSolutionsLink and remove /base directory
